% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS4Classes.R
\name{getS4Classes}
\alias{getS4Classes}
\title{Helper: Get S4-classes defined in a set of packages}
\usage{
getS4Classes(packages)
}
\arguments{
\item{packages}{character: R packages}
}
\value{
A data.frame (Superclasses is across packages and Subclasses is
  within a package)
}
\description{
This function uses getClasses to get S4-classes from packages and
isVirtualClass and isClassUnion to describe them.
}
\examples{
# Single package
getS4Classes("S4Vectors")

# Multiple packages
getS4Classes(c("IRanges", "GenomicRanges"))
}
