% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getS4Inheritance.R
\name{getS4Inheritance}
\alias{getS4Inheritance}
\title{Helper: Get S4 inheritance for a set of S4 classes}
\usage{
getS4Inheritance(S4Classes, packages)
}
\arguments{
\item{S4Classes}{character: S4-classes}

\item{packages}{character: R-package of each S4-class}
}
\value{
data.frame
}
\description{
This function uses getClassDef to extract inheritance information between S4
classes.
}
\examples{
# Single package
one_package <- getS4Classes("GenomicRanges")
getS4Inheritance(S4Classes = one_package$S4Class,
                  packages = one_package$Package)

# Several packages
two_packages <- getS4Classes(c("IRanges", "GenomicRanges"))
getS4Inheritance(S4Classes = two_packages$S4Class,
                  packages = two_packages$Package)
}
