% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotS4ClassGraph.R
\name{plotS4ClassGraph}
\alias{plotS4ClassGraph}
\title{Plot S4 classes and inheritance as a graph}
\usage{
plotS4ClassGraph(packages, plot = TRUE)
}
\arguments{
\item{packages}{character: R-packages}

\item{plot}{logical: Whether to plot (TRUE) or just return the underlying
plot data (FALSE)}
}
\value{
ggplot if plot=TRUE or list if plot=FALSE
}
\description{
This function plots all S4-classes from one or more R-packages as graph,
where each node is an S4 class and directed edges represents inheritance.
Virtual classes are indicated via node shapes. The graph stops at S4 defined
outside the set or R-packages in question.
}
\examples{
# Single package
plotS4ClassGraph("GenomicRanges")

# Several packages:
plotS4ClassGraph(c("IRanges", "GenomicRanges"))
}
