% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Cartographer-package.R
\docType{package}
\name{S4Cartographer-package}
\alias{S4Cartographer}
\alias{S4Cartographer-package}
\title{S4Cartographer: Visualize S4 class inheritance across R/Bioconductor packages}
\description{
S4 Object-Oriented Programming is fundamental to the Bioconductor ecosystem, enabling modularity and code reuse across hundreds of packages. However, the multi-layered inheritance structures within and between packages can be difficult to navigate. S4Cartographer facilitates the exploration of these relationships by visualizing S4 class hierarchies as directed graphs. By representing classes as nodes and inheritance as edges, it provides a clear overview of complex dependency networks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MalteThodberg/S4Cartographer}
  \item Report bugs at \url{https://github.com/MalteThodberg/S4Cartographer/issues}
}

}
\author{
\strong{Maintainer}: Malte Thodberg \email{maltethodberg@gmail.com} (\href{https://orcid.org/0000-0001-6244-3841}{ORCID})

}
\keyword{internal}
