% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/S4Cartographer-package.R
\docType{package}
\name{S4Cartographer-package}
\alias{S4Cartographer}
\alias{S4Cartographer-package}
\title{S4Cartographer: Visualize S4 class inheritance across R/Bioconductor packages}
\description{
Visualize S4 class inheritance across R/Bioconductor packages by plotting them as a graph, where each node is an S4 class and edges represent inheritance.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/MalteThodberg/S4Cartographer}
  \item Report bugs at \url{https://github.com/MalteThodberg/S4Cartographer/issues}
}

}
\author{
\strong{Maintainer}: Malte Thodberg \email{maltethodberg@gmail.com} (\href{https://orcid.org/0000-0001-6244-3841}{ORCID})

}
\keyword{internal}
